+++ raiden-ts/CHANGELOG.md
@@ -10,7 +10,11 @@
 ### Changed
 - [#2409] Lower default payment expiration to 1.1 × reveal timeout
 
+### Fixed
+- [#2352] Presence bug, transport fixes and performance improvements
+
 [#211]: https://github.com/raiden-network/light-client/issues/211
+[#2352]: https://github.com/raiden-network/light-client/issues/2352
 [#2409]: https://github.com/raiden-network/light-client/issues/2409
 [#2417]: https://github.com/raiden-network/light-client/pull/2417
 [#2444]: https://github.com/raiden-network/light-client/issues/2444